+++ src/aks-preview/azext_aks_preview/_params.py
@@ -23,6 +23,7 @@ from azure.cli.command_modules.acs._validators import (
     validate_nat_gateway_idle_timeout,
     validate_nat_gateway_managed_outbound_ip_count,
 )
+# from azure.cli.core.api import get_config_dir
 from azure.cli.core.commands.parameters import (
     edge_zone_type,
     file_type,
@@ -223,6 +224,7 @@ from azext_aks_preview._validators import (
     validate_max_blocked_nodes,
     validate_resource_group_parameter,
     validate_location_resource_group_cluster_parameters,
+    # validate_agent_config_file,
 )
 from azext_aks_preview.azurecontainerstorage._consts import (
     CONST_ACSTOR_ALL,
@@ -862,6 +864,7 @@ def load_arguments(self, _):
             ),
         )
         c.argument("dns_zone_resource_ids", is_preview=True)
+        c.argument('disable_run_command', action='store_true')
         c.argument("enable_keda", action="store_true", is_preview=True)
         c.argument(
             "enable_vpa",
@@ -1073,6 +1076,7 @@ def load_arguments(self, _):
         c.argument("vm_sizes", is_preview=True)
         c.argument("enable_imds_restriction", action="store_true", is_preview=True)
         c.argument("enable_managed_system_pool", action="store_true", is_preview=True)
+        c.argument("enable_upstream_kubescheduler_user_configuration", action="store_true", is_preview=True)
 
     with self.argument_context("aks update") as c:
         # managed cluster paramerters
@@ -1302,6 +1306,8 @@ def load_arguments(self, _):
             validator=validate_apiserver_subnet_id,
             is_preview=True,
         )
+        c.argument('enable_run_command', action='store_true')
+        c.argument('disable_run_command', action='store_true')
         c.argument("enable_keda", action="store_true", is_preview=True)
         c.argument("disable_keda", action="store_true", is_preview=True)
         c.argument(
@@ -1516,6 +1522,8 @@ def load_arguments(self, _):
         c.argument('migrate_vmas_to_vms', is_preview=True, action='store_true')
         c.argument("disable_http_proxy", action="store_true", is_preview=True)
         c.argument("enable_http_proxy", action="store_true", is_preview=True)
+        c.argument("enable_upstream_kubescheduler_user_configuration", action="store_true", is_preview=True)
+        c.argument("disable_upstream_kubescheduler_user_configuration", action="store_true", is_preview=True)
 
     with self.argument_context("aks upgrade") as c:
         c.argument("kubernetes_version", completer=get_k8s_upgrades_completion_list)
@@ -1772,6 +1780,11 @@ def load_arguments(self, _):
         )
         # virtual machines
         c.argument("vm_sizes", is_preview=True)
+        # local DNS
+        c.argument(
+            'localdns_config',
+            help='Path to a JSON file to configure the local DNS profile for a new nodepool.'
+        )
 
     with self.argument_context("aks nodepool update") as c:
         c.argument(
@@ -1863,6 +1876,11 @@ def load_arguments(self, _):
         c.argument(
             "disable_fips_image",
             action="store_true"
+        )
+        # local DNS
+        c.argument(
+            'localdns_config',
+            help='Path to a JSON file to configure the local DNS profile for an existing nodepool.',
         )
 
     with self.argument_context("aks nodepool upgrade") as c:
@@ -2750,6 +2768,82 @@ def load_arguments(self, _):
                 options_list=["--name", "-n"],
                 help="Name of the load balancer configuration. Required.",
             )
+
+    with self.argument_context("aks bastion") as c:
+        c.argument("bastion")
+        c.argument("port", type=int)
+        c.argument("admin", action="store_true")
+        c.argument(
+            "yes",
+            options_list=["--yes", "-y"],
+            help="Do not prompt for confirmation.",
+            action="store_true",
+        )
+
+# pylint: disable=line-too-long
+#     with self.argument_context("aks agent") as c:
+#         c.positional(
+#             "prompt",
+#             help="Ask any question and answer using available tools.",
+#         )
+#         c.argument(
+#             "resource_group_name",
+#             options_list=["--resource-group", "-g"],
+#             help="Name of resource group.",
+#             required=False,
+#         )
+#         c.argument(
+#             "name",
+#             options_list=["--name", "-n"],
+#             help="Name of the managed cluster.",
+#             required=False,
+#         )
+#         c.argument(
+#             "max_steps",
+#             type=int,
+#             default=10,
+#             required=False,
+#             help="Maximum number of steps the LLM can take to investigate the issue.",
+#         )
+#         c.argument(
+#             "config_file",
+#             default=os.path.join(get_config_dir(), "aksAgent.config"),
+#             validator=validate_agent_config_file,
+#             required=False,
+#             help="Path to the config file.",
+#         )
+#         c.argument(
+#             "model",
+#             help="The model to use for the LLM.",
+#             required=False,
+#             type=str,
+#         )
+#         c.argument(
+#             "api-key",
+#             help="API key to use for the LLM (if not given, uses environment variables AZURE_API_KEY, OPENAI_API_KEY)",
+#             required=False,
+#             type=str,
+#         )
+#         c.argument(
+#             "no_interactive",
+#             help="Disable interactive mode. When set, the agent will not prompt for input and will run in batch mode.",
+#             action="store_true",
+#         )
+#         c.argument(
+#             "no_echo_request",
+#             help="Disable echoing back the question provided to AKS Agent in the output.",
+#             action="store_true",
+#         )
+#         c.argument(
+#             "show_tool_output",
+#             help="Show the output of each tool that was called.",
+#             action="store_true",
+#         )
+#         c.argument(
+#             "refresh_toolsets",
+#             help="Refresh the toolsets status.",
+#             action="store_true",
+#         )
 
 
 def _get_default_install_location(exe_name):