+++ src/containerapp/HISTORY.rst
@@ -4,8 +4,7 @@ Release History
 ===============
 upcoming
 ++++++
-* 'az containerapp update': Disallow changing `--revisions-mode` to Labels.
-* 'az containerapp up': Disallow changing `--revisions-mode` to Labels.
+* 'az containerapp update/up': Disallow changing `--revisions-mode` to Labels.
 * 'az containerapp session code-interpreter': Fix `--path` in examples
 
 1.2.0b3